Release 1.0.0-RC.2

+++ server/src/main/java/datart/server/base/dto/ScheduleBaseInfo.java
@@ -40,7 +40,7 @@ public class ScheduleBaseInfo {
 
     private Boolean isFolder;
 
-    private Integer index;
+    private Double index;
 
 
     public ScheduleBaseInfo(Schedule schedule) {
